split benchmarks
