+++ Cargo.toml
@@ -24,7 +24,11 @@ csv = "1.3.0"
 rstest = "0.23.0"
 
 [[bench]]
-name = "storage_bench"
+name = "random_data_bench"
+harness = false
+
+[[bench]]
+name = "real_data_bench"
 harness = false
 
 [profile.bench]
